src/main/java/csci310/servlets/DashboardServlet.java
package csci310.servlets;

import javax.servlet.annotation.WebServlet;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

/*
 * Main servlet for dashboard page that contains instances of other servlets
 */
@WebServlet("/dashboard")
public class DashboardServlet extends HttpServlet {
	private static final long serialVersionUID = 1L;
	
	@Override
	protected void doGet(HttpServletRequest request, HttpServletResponse response) {
		return;
	}
	
	@Override
	protected void doPost(HttpServletRequest request, HttpServletResponse response) {
		return;
	}
	
}
